+++ packages/flutter_deck/example/lib/shortcuts/shortcuts.dart
@@ -1,0 +1,1 @@
+export 'skip_slide_shortcut.dart';