+++ packages/flutter_deck/lib/src/controls/controls.dart
@@ -2,5 +2,6 @@ export 'flutter_deck_controls.dart';
 export 'flutter_deck_controls_configuration.dart';
 export 'flutter_deck_controls_listener.dart';
 export 'flutter_deck_controls_notifier.dart';
+export 'flutter_deck_shortcut.dart';
 export 'fullscreen/fullscreen.dart';
 export 'l10n/l10n.dart';