+++ packages/flutter_deck/test/src/controls/flutter_deck_controls_configuration_test.dart
@@ -68,12 +68,14 @@ void main() {
       expect(configuration.previousSlide, const {SingleActivator(LogicalKeyboardKey.arrowLeft)});
       expect(configuration.toggleMarker, const {SingleActivator(LogicalKeyboardKey.keyM)});
       expect(configuration.toggleNavigationDrawer, const {SingleActivator(LogicalKeyboardKey.period)});
+      expect(configuration.customShortcuts, isEmpty);
     });
 
     test('disabled factory should create disabled configuration', () {
       const configuration = FlutterDeckShortcutsConfiguration.disabled();
 
       expect(configuration.enabled, false);
+      expect(configuration.customShortcuts, isEmpty);
     });
   });
 }